+++ R/depigner-package.R
@@ -4,6 +4,7 @@
 #' Biostatistics, Epidemiology and Public Health of the Department of Cardiac,
 #' Thoracic and Vascular sciences at the university of Padova.
 #'
+#' @importFrom tibble tibble
 #' @docType package
 #' @keywords internal
 "_PACKAGE"